Collapse unused app and user POI interfaces
